+++ sdk/resourcemanager/sqlvirtualmachine/armsqlvirtualmachine/zz_generated_constants.go
@@ -10,7 +10,7 @@ package armsqlvirtualmachine
 
 const (
 	moduleName    = "armsqlvirtualmachine"
-	moduleVersion = "v0.3.0"
+	moduleVersion = "v0.4.0"
 )
 
 // BackupScheduleType - Backup schedule type.